requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
pyperclip>=1.8.0
psutil>=5.9.0
